+++ psi4/src/psi4/libciomr/block_matrix.cc
@@ -60,8 +60,8 @@ namespace psi {
 ** Allocates memory for an n x m matrix and returns a pointer to the
 ** first row.
 **
-** \param n = number of rows (unsigned long to allow large matrices)
-** \param m = number of columns (unsigned long to allow large matrices)
+** \param n = number of rows (size_t to allow large matrices)
+** \param m = number of columns (size_t to allow large matrices)
 ** \param memlock = optional bool indicating whether to lock memory
 **   into physical RAM or not, and available only where _POSIX_MEMLOCK
 **   is defined. Defaults to false if not specified.
@@ -106,11 +106,11 @@ double ** block_matrix(size_t n, size_t m, bool memlock)
     if (memlock) {
 
         char* addr = (char*) B;
-        unsigned long size = m*n*(unsigned long)sizeof(double);
-        unsigned long page_offset, page_size;
+        size_t size = m*n*(size_t)sizeof(double);
+        size_t page_offset, page_size;
 
         page_size = sysconf(_SC_PAGESIZE);
-        page_offset = (unsigned long) addr % page_size;
+        page_offset = (size_t) addr % page_size;
 
         addr -= page_offset;  /* Adjust addr to page boundary */
         size += page_offset;  /* Adjust size with page_offset */
@@ -122,9 +122,9 @@ double ** block_matrix(size_t n, size_t m, bool memlock)
         }
 
         addr = (char*) A;
-        size = n*(unsigned long)sizeof(double*);
+        size = n*(size_t)sizeof(double*);
 
-        page_offset = (unsigned long) addr % page_size;
+        page_offset = (size_t) addr % page_size;
 
         addr -= page_offset;  /* Adjust addr to page boundary */
         size += page_offset;  /* Adjust size with page_offset */
